documentation: move all popups to popupfolder
