+++ frontend/src/components/popups/AddPlacePopup.js
