+++ frontend/src/components/popups/ImagePopup.js
